Check_points_in_domain respects multi-column encoding and multiple points
Categorical parameters can encode it's domain as multiple continuous
parameters. Therefore the feature-space is not necessary equal to the
number of parameters. A slice of the input features has to be passed
to the parameter's check in domain method.

Parameter.check_in_domain responds a single boolean for all of the input points.
Therefore it has to be called once per point per parameter